Extracted DataFrame from ETL class
